requirements.txt
aiohttp==3.12.12
anthropic==0.59.0
google-auth==2.40.3
google-genai==1.19.0
elevenlabs==2.16.0
python-dotenv==1.1.1
pillow~=11.3.0
protobuf~=6.32.0
python-telegram-bot~=22.4